refactor luigi config
