+++ core/src/luigi-config/luigi-config.js
@@ -1,379 +1,20 @@
-import LuigiClient from '@kyma-project/luigi-client';
 
-import convertToNavigationTree from './microfrontend-converter';
 import {
-  hideDisabledNodes,
-  getSystemNamespaces,
-  shouldShowSystemNamespaces,
   saveCurrentLocation,
-  getPreviousLocation
-} from './navigation-helpers';
+  getPreviousLocation,
+  setLimitExceededErrorsMessages
+} from './navigation/navigation-helpers';
 import { communication } from './communication';
-import { CONSOLE_INIT_DATA, GET_MICROFRONTENDS } from './queries';
-import navigationPermissionChecker, {
-  setInitValues,
-  backendModules
-} from './permissions';
+import { config } from './config';
+import { navigation, getNavigationData, getToken } from './navigation/navigation-data-init';
 
-var clusterConfig = window['clusterConfig'] || INJECTED_CLUSTER_CONFIG;
-var k8sDomain = (clusterConfig && clusterConfig['domain']) || 'kyma.local';
 
-var k8sServerUrl = 'https://apiserver.' + k8sDomain;
-
-export const config = {
-  domain: 'kyma.local',
-  localDomain: 'console-dev.kyma.local',
-  serviceCatalogModuleUrl: 'https://catalog.' + k8sDomain,
-  serviceInstancesModuleUrl: 'https://instances.' + k8sDomain,
-  lambdasModuleUrl: 'https://lambdas-ui.' + k8sDomain,
-  serviceBrokersModuleUrl: 'https://brokers.' + k8sDomain,
-  docsModuleUrl: 'https://docs.' + k8sDomain,
-  addOnsModuleUrl: 'https://addons.' + k8sDomain,
-  logsModuleUrl: 'https://log-ui.' + k8sDomain,
-  coreModuleUrl: 'https://core-ui.' + k8sDomain,
-  graphqlApiUrl: 'https://console-backend.' + k8sDomain + '/graphql',
-  disabledNavigationNodes: '',
-  systemNamespaces:
-    'compass-system istio-system knative-eventing knative-serving kube-public kube-system kyma-backup kyma-installer kyma-integration kyma-system natss kube-node-lease kubernetes-dashboard serverless-system',
-  compassDefaultTenant: '',
-  ...clusterConfig
-};
-
-var token;
-if (sessionStorage.getItem('luigi.auth')) {
-  token = JSON.parse(sessionStorage.getItem('luigi.auth')).idToken;
+function getFreshKeys() {
+  // manually re-fetching keys, since this is a major pain point
+  // until dex has possibility of no-cache
+  return fetch('https://dex.' + config.domain + '/keys', { cache: 'no-cache' });
 }
 
-const consoleViewGroupName = '_console_';
-const coreUIViewGroupName = '_core_ui_';
-const systemNamespaces = getSystemNamespaces(config.systemNamespaces);
-
-let navigation = {
-  viewGroupSettings: {
-    _console_: {
-      preloadUrl: '/consoleapp.html#/home/preload'
-    },
-    _core_ui_: {
-      preloadUrl: config.coreModuleUrl + '/preload'
-    }
-  },
-  nodeAccessibilityResolver: navigationPermissionChecker,
-  contextSwitcher: {
-    defaultLabel: 'Select Namespace ...',
-    parentNodePath: '/home/namespaces', // absolute path
-    lazyloadOptions: true, // load options on click instead on page load
-    options: getNamespaces,
-    actions: [
-      {
-        label: '+ New Namespace',
-        link: '/home/workspace?~showModal=true'
-      }
-    ]
-  },
-  profile: {
-    items: [
-      {
-        icon: 'settings',
-        label: 'Settings',
-        link: '/home/settings'
-      }
-    ]
-  }
-};
-
-function getNodes(context) {
-  var namespace = context.namespaceId;
-  var staticNodes = [
-    {
-      link: '/home/workspace',
-      label: 'Back to Namespaces',
-      icon: 'nav-back'
-    },
-    {
-      pathSegment: 'details',
-      label: 'Overview',
-      viewUrl: '/consoleapp.html#/home/namespaces/' + namespace + '/details',
-      icon: 'product'
-    },
-    {
-      category: { label: 'Service Management', icon: 'add-coursebook' },
-      pathSegment: '_service_management_category_placeholder_',
-      hideFromNav: true
-    },
-    {
-      category: { label: 'Configuration', icon: 'key-user-settings' },
-      pathSegment: '_configuration_category_placeholder_',
-      hideFromNav: true
-    },
-    {
-      category: 'Configuration',
-      pathSegment: 'permissions',
-      navigationContext: 'permissions',
-      label: 'Permissions',
-      viewUrl:
-        '/consoleapp.html#/home/namespaces/' + namespace + '/permissions',
-      keepSelectedForChildren: true,
-      children: [
-        {
-          pathSegment: 'roles',
-          children: [
-            {
-              pathSegment: ':name',
-              viewUrl:
-                '/consoleapp.html#/home/namespaces/' +
-                namespace +
-                '/permissions/roles/:name'
-            }
-          ]
-        }
-      ]
-    },
-    {
-      category: 'Configuration',
-      pathSegment: 'resources',
-      navigationContext: 'resources',
-      label: 'Resources',
-      viewUrl: '/consoleapp.html#/home/namespaces/' + namespace + '/resources'
-    },
-    {
-      category: 'Configuration',
-      pathSegment: 'config-maps',
-      navigationContext: 'config-maps',
-      label: 'Config maps',
-      viewUrl: '/consoleapp.html#/home/namespaces/' + namespace + '/configmaps'
-    },
-    {
-      category: { label: 'Development', icon: 'source-code' },
-      pathSegment: '_development_category_placeholder_',
-      hideFromNav: true
-    },
-    {
-      category: { label: 'Operation', icon: 'instance' },
-      pathSegment: 'deployments',
-      navigationContext: 'deployments',
-      label: 'Deployments',
-      viewUrl: '/consoleapp.html#/home/namespaces/' + namespace + '/deployments'
-    },
-    {
-      category: 'Operation',
-      pathSegment: 'replica-sets',
-      navigationContext: 'replica-sets',
-      label: 'Replica Sets',
-      viewUrl: '/consoleapp.html#/home/namespaces/' + namespace + '/replicaSets'
-    },
-    {
-      category: 'Operation',
-      pathSegment: 'pods',
-      navigationContext: 'pods',
-      label: 'Pods',
-      viewUrl: '/consoleapp.html#/home/namespaces/' + namespace + '/pods'
-    },
-    {
-      category: 'Operation',
-      pathSegment: 'services',
-      navigationContext: 'services',
-      label: 'Services',
-      viewUrl: '/consoleapp.html#/home/namespaces/' + namespace + '/services',
-      keepSelectedForChildren: true,
-      children: [
-        {
-          pathSegment: 'details',
-          children: [
-            {
-              pathSegment: ':name',
-              viewUrl:
-                '/consoleapp.html#/home/namespaces/' +
-                namespace +
-                '/services/:name',
-              children: [
-                {
-                  pathSegment: 'apis',
-                  children: [
-                    {
-                      pathSegment: 'create',
-                      viewUrl:
-                        '/consoleapp.html#/home/namespaces/' +
-                        namespace +
-                        '/services/:name/apis/create'
-                    },
-                    {
-                      pathSegment: 'details',
-                      children: [
-                        {
-                          pathSegment: ':apiName',
-                          viewUrl:
-                            '/consoleapp.html#/home/namespaces/' +
-                            namespace +
-                            '/services/:name/apis/details/:apiName'
-                        }
-                      ]
-                    }
-                  ]
-                }
-              ]
-            }
-          ]
-        }
-      ]
-    },
-    {
-      category: 'Operation',
-      pathSegment: 'secrets',
-      navigationContext: 'secrets',
-      label: 'Secrets',
-      viewUrl: '/consoleapp.html#/home/namespaces/' + namespace + '/secrets',
-      keepSelectedForChildren: true,
-      children: [
-        {
-          pathSegment: 'details',
-          children: [
-            {
-              pathSegment: ':name',
-              viewUrl:
-                '/consoleapp.html#/home/namespaces/' +
-                namespace +
-                '/secrets/:name'
-            }
-          ]
-        }
-      ]
-    },
-    {
-      category: { label: 'Experimental', icon: 'lab' },
-      hideFromNav: true
-    }
-  ];
-  return Promise.all([
-    getMicrofrontends(namespace),
-    Promise.resolve(window.clusterMicrofrontendNodesForNamespace)
-  ])
-    .then(function(values) {
-      var nodeTree = [...staticNodes];
-      values.forEach(function(val) {
-        nodeTree = [].concat.apply(nodeTree, val);
-      });
-
-      hideDisabledNodes(config.disabledNavigationNodes, nodeTree, true);
-      return nodeTree;
-    })
-    .catch(err => {
-      const errParsed = JSON.parse(err);
-      console.error('Error', errParsed);
-      const settings = {
-        text: `Namespace ${errParsed.details.name} not found.`,
-        type: 'error'
-      };
-      LuigiClient.uxManager().showAlert(settings);
-    });
-}
-
-/**
- * getMicrofrontends
- * @param {string} namespace k8s namespace name
- */
-const getMicrofrontends = async namespace => {
-  const segmentPrefix = 'mf-';
-
-  const cacheName = '_console_mf_cache_';
-  if (!window[cacheName]) {
-    window[cacheName] = {};
-  }
-  const cache = window[cacheName];
-  const cacheKey = segmentPrefix + namespace;
-  const fromCache = cache[cacheKey];
-
-  return (
-    fromCache ||
-    fetchFromGraphQL(GET_MICROFRONTENDS, { namespace }, true)
-      .then(result => {
-        if (!result.microFrontends || !result.microFrontends.length) {
-          return [];
-        }
-        return result.microFrontends.map(function(item) {
-          if (item.navigationNodes) {
-            return convertToNavigationTree(
-              item.name,
-              item,
-              config,
-              navigation,
-              consoleViewGroupName,
-              segmentPrefix
-            );
-          }
-          return [];
-        });
-      })
-      .catch(err => {
-        console.error(`Error fetching Microfrontend ${name}: ${err}`);
-        return [];
-      })
-      .then(result => {
-        cache[cacheKey] = new Promise(function(resolve) {
-          resolve(result);
-        });
-        return result;
-      })
-  );
-};
-
-function fetchFromKyma(url) {
-  return new Promise(function(resolve, reject) {
-    var xmlHttp = new XMLHttpRequest();
-    xmlHttp.onreadystatechange = function() {
-      if (xmlHttp.readyState == 4 && xmlHttp.status == 200) {
-        resolve(JSON.parse(xmlHttp.response));
-      } else if (xmlHttp.readyState == 4 && xmlHttp.status != 200) {
-        if (xmlHttp.status === 401) {
-          relogin();
-        }
-        reject(xmlHttp.response);
-      }
-    };
-
-    xmlHttp.open('GET', url, true);
-    xmlHttp.setRequestHeader('Authorization', 'Bearer ' + token);
-    xmlHttp.send(null);
-  });
-}
-
-function fetchFromGraphQL(query, variables, gracefully) {
-  return new Promise(function(resolve, reject) {
-    var xmlHttp = new XMLHttpRequest();
-    xmlHttp.onreadystatechange = function() {
-      if (xmlHttp.readyState == 4 && xmlHttp.status == 200) {
-        try {
-          const response = JSON.parse(xmlHttp.response);
-          if (response && response.errors) {
-            reject(response.errors[0].message);
-          } else if (response && response.data) {
-            return resolve(response.data);
-          }
-          resolve(response);
-        } catch {
-          reject(xmlHttp.response);
-        }
-      } else if (xmlHttp.readyState == 4 && xmlHttp.status != 200) {
-        if (xmlHttp.status === 401) {
-          relogin();
-        }
-        if (!gracefully) {
-          reject(xmlHttp.response);
-        } else {
-          resolve(null);
-        }
-      }
-    };
-
-    xmlHttp.open('POST', config.graphqlApiUrl, true);
-    xmlHttp.setRequestHeader('Authorization', 'Bearer ' + token);
-    xmlHttp.setRequestHeader('Content-Type', 'application/json');
-    xmlHttp.send(JSON.stringify({ query, variables }));
-  });
-}
-
-function getConsoleInitData() {
-  return fetchFromGraphQL(CONSOLE_INIT_DATA, undefined, true);
-}
 
 window.addEventListener('message', e => {
   const SHOW_SYSTEM_NAMESPACES_CHANGE_EVENT =
@@ -389,335 +30,86 @@ window.addEventListener('message', e => {
   }
 });
 
-function getNamespaces() {
-  if (window.Luigi.cachedNamespaces) {
-    return createNamespacesList(window.Luigi.cachedNamespaces);
-  } else {
-    return fetchFromKyma(k8sServerUrl + '/api/v1/namespaces')
-      .then(response => {
-        window.Luigi.cachedNamespaces = response.items;
-        return createNamespacesList(response.items);
-      })
-      .catch(function catchNamespaces(err) {
-        window.Luigi.cachedNamespaces = null;
-        console.error('get namespace: error', err);
-      });
-  }
-}
 
-function createNamespacesList(rawNamespaces) {
-  var namespaces = [];
-  rawNamespaces.map(namespace => {
-    if (namespace.status && namespace.status.phase !== 'Active') {
-      return; //"pretend" that inactive namespace is already removed
-    }
-    const namespaceName = namespace.metadata.name;
-    const isSystemNamespace = systemNamespaces.some(namespace => {
-      return namespace === namespaceName;
-    });
-    const alternativeLocation = getCorrespondingNamespaceLocation(
-      namespaceName
-    );
-    if (!isSystemNamespace || shouldShowSystemNamespaces()) {
-      namespaces.push({
-        category: 'Namespaces',
-        label: namespaceName,
-        pathValue: alternativeLocation || namespaceName
-      });
-    }
-  });
-  return namespaces;
-}
+getFreshKeys()
 
-function getCorrespondingNamespaceLocation(namespaceName) {
-  const addressTokens = window.location.pathname.split('/');
-  // check if we are in namespaces context
-  if (addressTokens[2] !== 'namespaces') {
-    return null;
-  }
-  // check if any path after namespace name exists - if not,
-  // it will default to namespace name (and then to '/details')
-  if (!addressTokens[4]) {
-    return null;
-  }
-  return namespaceName + '/' + addressTokens.slice(4).join('/');
-}
-
-function relogin() {
-  saveCurrentLocation();
-  sessionStorage.removeItem('luigi.auth');
-  location.reload();
-}
-
-function getFreshKeys() {
-  // manually re-fetching keys, since this is a major pain point
-  // until dex has possibility of no-cache
-  return fetch('https://dex.' + k8sDomain + '/keys', { cache: 'no-cache' });
-}
-
-let clusterMicrofrontendNodes = [];
-var initPromises = [getFreshKeys()];
-let kymaVersion;
-
-if (token) {
-  initPromises.push(getConsoleInitData());
-} else {
-  saveCurrentLocation();
-}
-
-Promise.all(initPromises)
-  .then(
-    res => {
-      if (token && res[1]) {
-        const modules = res[1].backendModules;
-        const subjectRules = res[1].selfSubjectRules;
-        const cmfs = res[1].clusterMicroFrontends;
-        kymaVersion = `Kyma version: ${res[1].versionInfo.kymaVersion}`;
-        setInitValues(
-          (modules && modules.map(m => m.name)) || [],
-          subjectRules || []
-        );
-
-        if (cmfs && cmfs.length > 0) {
-          clusterMicrofrontendNodes = cmfs
-            .filter(cmf => cmf.placement === 'cluster')
-            .map(cmf => {
-              if (cmf.navigationNodes) {
-                var tree = convertToNavigationTree(
-                  cmf.name,
-                  cmf,
-                  config,
-                  navigation,
-                  consoleViewGroupName,
-                  'cmf-'
-                );
-                return tree;
-              }
-              return [];
-            });
-          window.clusterMicrofrontendNodesForNamespace = cmfs
-            .filter(
-              cmf =>
-                cmf.placement === 'namespace' || cmf.placement === 'environment'
-            )
-            .map(cmf => {
-              // console.log(cmf.name, cmf);
-              if (cmf.navigationNodes) {
-                return convertToNavigationTree(
-                  cmf.name,
-                  cmf,
-                  config,
-                  navigation,
-                  consoleViewGroupName,
-                  'cmf-'
-                );
-              }
-              return [];
-            });
-        }
-      }
+const luigiConfig = {
+  auth: {
+    use: 'openIdConnect',
+    openIdConnect: {
+      authority: 'https://dex.' + config.domain,
+      client_id: 'console',
+      scope:
+        'audience:server:client_id:kyma-client audience:server:client_id:console openid email profile groups',
+      automaticSilentRenew: true,
+      loadUserInfo: false,
+      logoutUrl: 'logout.html'
     },
-    err => {
-      // console.error(err);
-    }
-  )
-  // 'Finally' not supported by IE and FIREFOX (if 'finally' is needed, update your .babelrc)
-  .then(() => {
-    (navigation.nodes = () => [
-      {
-        pathSegment: 'home',
-        hideFromNav: true,
-        context: {
-          idToken: token,
-          backendModules,
-          systemNamespaces,
-          tenantId: config.compassDefaultTenant,
-          showSystemNamespaces:
-          localStorage.getItem('console.showSystemNamespaces') === 'true'
-        },
-        children: function() {
-          var staticNodes = [
-            {
-              pathSegment: 'workspace',
-              label: 'Namespaces',
-              viewUrl: config.coreModuleUrl + '/namespaces',
-              icon: 'dimension',
-              viewGroup: coreUIViewGroupName
-            },
-            {
-              pathSegment: 'namespaces',
-              viewUrl: '/consoleapp.html#/home/namespaces/workspace',
-              hideFromNav: true,
-              viewGroup: consoleViewGroupName,
-              children: [
-                {
-                  pathSegment: ':namespaceId',
-                  context: {
-                    environmentId: ':namespaceId',
-                    namespaceId: ':namespaceId'
-                  },
-                  children: getNodes,
-                  navigationContext: 'namespaces',
-                  defaultChildNode: 'details'
-                }
-              ]
-            },
-            {
-              category: { label: 'Integration', icon: 'overview-chart' },
-              pathSegment: '_integration_category_placeholder_',
-              hideFromNav: true
-            },
-            {
-              pathSegment: 'settings',
-              navigationContext: 'settings',
-              label: 'General Settings',
-              category: { label: 'Settings', icon: 'settings' },
-              viewUrl: '/consoleapp.html#/home/settings/organisation',
-              viewGroup: consoleViewGroupName
-            },
-            {
-              pathSegment: 'global-permissions',
-              navigationContext: 'global-permissions',
-              label: 'Global Permissions',
-              category: 'Settings',
-              viewUrl: '/consoleapp.html#/home/settings/globalPermissions',
-              keepSelectedForChildren: true,
-              viewGroup: consoleViewGroupName,
-              children: [
-                {
-                  pathSegment: 'roles',
-                  children: [
-                    {
-                      pathSegment: ':name',
-                      viewUrl:
-                        '/consoleapp.html#/home/settings/globalPermissions/roles/:name'
-                    }
-                  ]
-                }
-              ],
-              requiredPermissions: [
-                {
-                  apiGroup: 'rbac.authorization.k8s.io',
-                  resource: 'clusterrolebindings',
-                  verbs: ['create']
-                }
-              ]
-            },
-            {
-              category: {
-                label: 'Diagnostics',
-                icon: 'electrocardiogram'
-              },
-              pathSegment: '_integration_category_placeholder_',
-              hideFromNav: true
-            },
-            {
-              category: { label: 'Experimental', icon: 'lab' },
-              hideFromNav: true
-            }
-          ];
-          const fetchedNodes = [].concat(...clusterMicrofrontendNodes);
-          const nodeTree = [...staticNodes, ...fetchedNodes];
-          hideDisabledNodes(config.disabledNavigationNodes, nodeTree, false);
-          return nodeTree;
+
+    events: {
+      onLogout: () => {
+        console.log('onLogout');
+      },
+      onAuthSuccessful: () => {
+        const prevLocation = getPreviousLocation();
+        if (prevLocation) {
+          window.location.replace(prevLocation);
         }
       },
-      {
-        pathSegment: 'docs',
-        viewUrl: config.docsModuleUrl,
-        label: 'Docs',
-        hideSideNav: true,
-        context: {
-          idToken: token,
-          backendModules
-        },
-        icon: 'sys-help',
-        children: [
-          {
-            pathSegment: ':group',
-            viewUrl: config.docsModuleUrl,
-            hideSideNav: true,
-            context: {
-              group: ':group'
-            },
-            children: [
-              {
-                pathSegment: ':topic',
-                viewUrl: config.docsModuleUrl,
-                hideSideNav: true,
-                context: {
-                  group: ':group',
-                  topic: ':topic'
-                }
-              }
-            ]
-          }
-        ]
+      onAuthExpired: () => {
+        console.log('onAuthExpired');
+      },
+      // TODO: define luigi-client api for getting errors
+      onAuthError: err => {
+        console.log('authErrorHandler 1', err);
       }
-    ]),
-      Luigi.setConfig({
-        auth: {
-          use: 'openIdConnect',
-          openIdConnect: {
-            authority: 'https://dex.' + k8sDomain,
-            client_id: 'console',
-            scope:
-              'audience:server:client_id:kyma-client audience:server:client_id:console openid profile email groups',
-            automaticSilentRenew: true,
-            loadUserInfo: false,
-            logoutUrl: 'logout.html'
-          },
-
-          events: {
-            onLogout: () => {
-              console.log('onLogout');
-            },
-            onAuthSuccessful: () => {
-              const prevLocation = getPreviousLocation();
-              if (prevLocation) {
-                window.location.replace(prevLocation);
-              }
-            },
-            onAuthExpired: () => {
-              console.log('onAuthExpired');
-            },
-            // TODO: define luigi-client api for getting errors
-            onAuthError: err => {
-              console.log('authErrorHandler 1', err);
-            }
-          },
-          storage: 'sessionStorage'
-        },
-        communication,
-        navigation,
-        routing: {
-          nodeParamPrefix: '~',
-          skipRoutingForUrlPatterns: [/access_token=/, /id_token=/]
-        },
-        settings: {
-          responsiveNavigation: 'simpleMobileOnly',
-          sideNavFooterText: kymaVersion,
-          header: () => {
-            const logo =
-              clusterConfig && clusterConfig.headerLogoUrl
-                ? clusterConfig.headerLogoUrl
-                : '/assets/logo.svg';
-            const title = clusterConfig ? clusterConfig.headerTitle : undefined;
-            const favicon = clusterConfig
-              ? clusterConfig.faviconUrl
-              : undefined;
-            return {
-              logo,
-              title,
-              favicon
-            };
+    },
+    storage: 'sessionStorage'
+  },
+  communication,
+  navigation,
+  routing: {
+    nodeParamPrefix: '~',
+    skipRoutingForUrlPatterns: [/access_token=/, /id_token=/]
+  },
+  settings: {
+    responsiveNavigation: 'simpleMobileOnly',
+    sideNavFooterText: '',
+    header: () => {
+      const logo =
+        config && config.headerLogoUrl
+          ? config.headerLogoUrl
+          : '/assets/logo.svg';
+      const title = config ? config.headerTitle : undefined;
+      const favicon = config
+        ? config.faviconUrl
+        : undefined;
+      return {
+        logo,
+        title,
+        favicon
+      };
+    }
+  },
+  lifecycleHooks: {
+    luigiAfterInit: () => {
+      const token = getToken()
+      if(token){
+        getNavigationData().then(
+          response => {
+            navigation.nodes = response[0];
+            luigiConfig.settings.sideNavFooterText = response[1];
+            Luigi.configChanged('navigation');
+            Luigi.configChanged('settings');
           }
-        }
-      });
-  })
-  .catch(err => {
-    console.error('Config Init Error', err);
-  });
+        )
+      } else {
+        saveCurrentLocation();
+      }
+    }
+  }
+};
+Luigi.setConfig(luigiConfig);
 
 window.addEventListener('message', e => {
   if (e.data.msg && e.data.msg === 'console.quotaexceeded') {
@@ -764,16 +156,4 @@ window.addEventListener('message', e => {
   }
 });
 
-function setLimitExceededErrorsMessages(limitExceededErrors) {
-  let limitExceededErrorscomposed = [];
-  limitExceededErrors.forEach(resource => {
-    if (resource.affectedResources && resource.affectedResources.length > 0) {
-      resource.affectedResources.forEach(affectedResource => {
-        limitExceededErrorscomposed.push(
-          `'${resource.resourceName}' by '${affectedResource}' (${resource.quotaName})`
-        );
-      });
-    }
-  });
-  return limitExceededErrorscomposed;
-}
+
