+++ core/src/luigi-config/navigation/permissions.js
@@ -1,5 +1,5 @@
 import rbacRulesMatched from './rbac-rules-matcher';
-import { config } from './luigi-config';
+import { config } from '../luigi-config';
 
 const ADMIN_ONLY_PATH_SEGMENTS = [
   'cmf-applications',
